Move autodiff vectors to LinearAlgebra folder
